Implemented legend and added more buttons
Still need to implement showing the solve algorithms (by stepping through solution) and saving them as a gif (look up python save as gif; will most likely use numpy module).
I was in the middle of allowing the user to interact with the maze: currently sets start and end nodes automatically; when the mouse is over a generated maze, it renders the 'currentNode'; when the maze is clicked on, it should save currentNode as activeNode (node that was selected), so then we could click "set start" or "set end", which would make the activeNode the start/end node.